Выполнил задание module5-task1 (разделение существующего кода на модули)
